Move ReadFile tests to config_test.go
